+++ Event.py
@@ -1,5 +1,4 @@
 import xml.etree.ElementTree as ET
-from decimal import Decimal
 
 class Event:
 	def __init__(self, Type, TimeOffset, EventParamTree):
@@ -8,8 +7,8 @@ class Event:
 		self.param = {}
 		if self.Type == 'Network':
 			self.param['Name'] = EventParamTree.find('Name').text
-			self.param['ClockAccuracy'] = Decimal(EventParamTree.find('ClockAccuracy').text)
-			self.param['PacketDeliveryRatio'] = Decimal(EventParamTree.find('PacketDeliveryRatio').text)
+			self.param['ClockAccuracy'] = float(EventParamTree.find('ClockAccuracy').text)
+			self.param['PacketDeliveryRatio'] = float(EventParamTree.find('PacketDeliveryRatio').text)
 
 	def __repr__(self):
 		string = '\n\tEvent of type ' + self.Type + ' occured at time ' + str(self.TimeOffset) + ' ms from start;'
